requirements/prod.txt
celery==5.3.6
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
psycopg2==2.9.9
requests==2.31.0
